Moved MapHeaderFlags to map_defs.h
