+++ src/object.cc
@@ -19,6 +19,7 @@
 #include "item.h"
 #include "light.h"
 #include "map.h"
+#include "map_defs.h"
 #include "memory.h"
 #include "party_member.h"
 #include "proto.h"
