+++ src/scripts.cc
@@ -25,6 +25,7 @@
 #include "game_mouse.h"
 #include "game_movie.h"
 #include "input.h"
+#include "map_defs.h"
 #include "memory.h"
 #include "message.h"
 #include "object.h"
